buck2: concurrency: only enter cleanup if the version has changed
Summary:
Now that we have first-class support for async cancellation in DICE, we don't
need to quarantine the full context as soon as we start cancelling.

Instead, we can keep reusing as long as the same version is active. If any
computations were cancelled, they will be handled individually in DICE, instead
of having the entire context quarantined.

If the version did update, we still need to enter the cleanup stage (this is
because DICE would attempt to run the same key at multiple versions otherwise).

To make this a little nicer, we allow skipping the cleanup step if the version
was not updated.

Reviewed By: bobyangyf

Differential Revision: D42885407

fbshipit-source-id: 3d5403c148f86eba1504e94d8e4375452d6d7020