Wait for symlink targets to materialize when materializing an entry in Windows
Summary:
To create symlinks in Windows, we need to know whether the target is a file or a directory. The symlink can be created without the target being present, but in our case we do not know the file type otherwise.

For Windows only, materialize the symlink targets first and then materialize the symlink.

Reviewed By: krallin

Differential Revision: D42970800

fbshipit-source-id: da7fda34fb64d20b925446d78de831e8d206490c