add project support
Summary:
Addresses one of the TODOs in D42969311:

> # 2. Finding the target's "project", which involves a rdeps search

The implementation uses a query to find all the `haskell_ide` targets in fbcode (the project universe), then grabs the first one that contains the target being solved, if any.

Note that the query to find all the `haskell_ide` targets is superexpensive and causes the Buck2 daemon to grow RSS to >50GB, so I have added support to provide the project universe externally.

Also, given that for Buck2 a `haskell_ide` target is nothing more than a set of targets, integration with Codehub projects should be trivial

Reviewed By: simonmar

Differential Revision: D42991514

fbshipit-source-id: 5093711fae83bc17bff8c09b96b1563ea503ae7a