buck2: dep files: don't reuse when the outputs changed
Summary:
If an action's command line didn't change, but the outputs list did, then you
currently get an error because we reuse a result that isn't satisfactory. This
diff fixes that by rejecting that dep file hit.

Ideally we'd just rehash them, but this isn't at the right layer to do that
(maybe it should, that's a bit debatable).

Reviewed By: IanChilds

Differential Revision: D43155065

fbshipit-source-id: 2c337a5f00edb1c5274ecbd776ab8b9f8954e61d