buck2: materializer: deferred: back out clean notifications stack
Summary:
Like it says in the title. I'm seeing reports of issues with missing files,
which is sort-of materializer related, and while it's not clear to me whether
they might be related to this stack of changes or not, or if they're
even more frequent than usual, reverting  seems like the safest approach,
so let's just do that for now.

I'll probably hold off on re-shipping this until we get all the concurrent
commands stuff rolled out so that we can trust that the rest of Buck2
isn't doing something wrong.

This reverts:

- D43003383 (https://github.com/facebookincubator/buck2/commit/94c61a7c6efd571b0ae50b571bce93ef5eaae432) buck2: materializer: deferred: notify the materializer when clean finishes
- D43000548 (https://github.com/facebookincubator/buck2/commit/888c850039a75bfbb98af0895898a3f81ce32d07) buck2: materializer: deferred: make the version API more robust
- D43000549 (https://github.com/facebookincubator/buck2/commit/9e6a0defe9557145ac8a654b0fc500ec950eb590) buck2: materializer: deferred: redo the version API
- D43000551 (https://github.com/facebookincubator/buck2/commit/0dafef719d09976f1098d9ed050934eda7d86f3b) buck2: materializer: deferred: add a type for Version
- D43000552 (https://github.com/facebookincubator/buck2/commit/b9aa19e3962668bf0fa38e190382b1600a0fb885) buck2: materializer: implement priorities
- D43000550 (https://github.com/facebookincubator/buck2/commit/204e605ebe5f90d33db0c03383f61a4c8a9a8d71) buck2: materializer: deferred: refactor the stream

Reviewed By: scottcao

Differential Revision: D43165680

fbshipit-source-id: 910fb7df39c87dadfa80dd7b04aa112788fd7fa2