Handle empty alias specially
Summary:
Empty cell alias in target name should always point to the current cell.

Enforce it with types: `CellAliasResolver` now conains a field: `current: CellName`, and cell mapping no longer contains and entry for empty string.

Also, in a lot of tests we do `CellName::unchecked_new("")`, while empty cell name should not be valid. But that fix is for another time.

Reviewed By: bobyangyf

Differential Revision: D43166434

fbshipit-source-id: 1abd07d21a7efde89e82a01af0e7fee9e425574c