Box a future
Summary:
This is seemingly a more minor change, but this affects the core future used
for all new dice tasks (and held alive for the duration of them).

Also, it seems like not all types are being printed by the tool I'm using as I
can't find the sizes of the futures that use this one. Using an existential
type + static assert shows that future goes from 624 -> 448.

```
< 168   impl futures::Future<Output = incremental::DidDepsChange> - dice/dice/src/incremental/mod.rs:900:10: 900:23
> 416   impl futures::Future<Output = incremental::DidDepsChange> - dice/dice/src/incremental/mod.rs:900:10: 900:23
```

Reviewed By: stepancheg

Differential Revision: D43143874

fbshipit-source-id: 05d830ef61e793eaae49251c383520dc15283206