Use "staged" ensure artifact for tsets
Summary:
A large build at peak can have nearly all of its actions (and tsets)
computations live waiting on their inputs (as buck2/dice is all pull
based, that's just how it works). The size of the futures (and data structures)
holding these edges that we are waiting on is critical.

This changes tset's calculation to hold only minimal data across the main yield
point where it is waiting for dependencies.

I didn't measure a peak change with this, but still seems appropriate.

Reviewed By: krallin

Differential Revision: D43149354

fbshipit-source-id: 917192a4017edfa5562f2acf51f97130c9705c52