Deduplicate .pyc files for inplace binaries
Summary:
In buck2 we pre-compile all sources in `python_library` targets in parallel. When an inplace binary produces the modules directory, these sources get merged and duplicates get filtered. This is simple for non-`.pyc` files: deduplication is based on the original source path; that's not enough for compiled files.

This diff relaxes the duplicate checking logic for `.pyc` files  to consider the first 16 bytes of them instead of their paths. Implementation based on https://fburl.com/code/nmd6x5fe

Reviewed By: BrandonTheBuilder

Differential Revision: D43193258

fbshipit-source-id: ba593da19049b2a991e4dc4577695d92fa4d0c30