buck2: move hidden components to Artifact
Summary:
Like it says in the title. I'd like this to be something we can set in
projected artifacts, and to do this we need to move it up one layer, from
BuildArtifact to Artifact.

Note: this changes the identifier for simple actions from short_path
to full path. That is a desirable change because we use it for the identifier,
which must be unique, so we should use the thing that is unique (but
in practice it won't matter because we never use artifacts with a prefix
with those actions).

Reviewed By: milend

Differential Revision: D43234441

fbshipit-source-id: 070de524bd83b1aecca220b78c96780ce6146f61