ctor instead of dice to inject BXL impl
Summary:
Got another idea how to do dependency injection in buck2.

Previously we did:
* DICE globals to propagate trait implementation, or
* pass the implementation manually where DICE is not available

That is not perfect (a lot to type).

Another idea is so simple, why I didn't think about it before. Let's use `static OnceCell` to store the implementation reference, and initialize it using `#[ctor]`.

This:
* removes some code
* (also makes invocation cheaper because no hashmap lookup)

This approach is not meant to be used if there's more than one trait implementation.

In particular, even if there's mock implementation of something which is only used in tests, and only one implementation in production, we should still not use this approach.

Reviewed By: krallin

Differential Revision: D43211395

fbshipit-source-id: b7da512a73de0dcc349b9c452a87f1ec4559acf8