Fix force-static heuristic
Summary:
Libraries will mark themselves as `preferred_linkage = "static"` to
ensure they always get linked into their dependents, which omnibus
honors in it's implementation: if a root node depends on a force-static
body node, that body node is statically linked into the root node; if
no other body node depends on that force-static node, then it will not
be added to the link line of the body.

Respect this heuristic in link groups, but always statically linking a force-
static node into it's dependent, even if it crosses a link group.

Reviewed By: christycylee

Differential Revision: D43222652

fbshipit-source-id: 89b3bf8d2e72ea2138e07d6bf9ca8c6161a3ca2d