buck2: directory: pass the hasher when fingerprinting
Summary:
Like it says in the title. The hasher used to be implicit (and specified as
a type parameter), but we need to change this because:

- We want to support keyed Blake3
- We want to make the type of hashes we generate configurable

So in order to do either of those things, we need to be able to inject
something that isn't static into the fingerprinting process.

Note that while we expect the fingerprinting to be consistent, it's not really
a requirement.

Reviewed By: stepancheg

Differential Revision: D43091982

fbshipit-source-id: bd4fd7bb9d55b4e110adf233cab679fca8c30449