buck2: thread a DigestConfig to SQLite -> materializer conversions
Summary:
Like it says in the title. We use this in the next diff. Note that the config
isn't *very* useful there (we use it to dedupe the empty digest to share it in
cache lookups, which might very well not be worth it here, but it's pretty easy
to just make this change and not think too hard about it).

Reviewed By: bobyangyf

Differential Revision: D43270644

fbshipit-source-id: 4a0ea3057f45daf34749038f87268d1c472c5e37