buck2: re_tree_to_directory: support querying multiple kinds of digests
Summary:
Like it says in the title. The way this method works is that it receives a
list of directories that reference each other by hash, but we don't receive
the hash.

So, if the directories reference each other by SHA1 hashes, we need to compute
SHA1 hashes of the directories. If the references are Blake3 hashes, we need
to compute those instead.

This code does of this, by allowing lookups to be done using the right kind of
hashes by having one map per hash.

Reviewed By: scottcao

Differential Revision: D43054883

fbshipit-source-id: 12df26eaca5623b683a0a970cace64ba448e5c3b