Only allow internal test runner to be used by default in open-source build
Summary:
Tpx should always be used instead of the internal test runner when running within Meta.

To enforce this, disable the default internal test runner rollback for non-open-source builds.

Note that it is still possible if absolutely necessary to use the internal runner in a non-open-source build, but it requires adding an additional wrapper script around the buck2 binary to make it work (since Buck2 expects test.v2_test_executor` to be a binary, not a command with arguments like `buck2 internal-test-runner`). This difficulty is a feature, not a bug!

Reviewed By: krallin

Differential Revision: D43163820

fbshipit-source-id: f0de396dff203b111393c9ec14c450e0730d80a0