Use config option to determine if android_instrumentation_test can run locally
Summary:
When we're running an `android_instrumentation_test` on Sandcastle, we _always_ want to do it using RE. I had erroneously thought that we could do that using `--unstable-force-tests-on-re`, but the "force" here doesn't force it to use RE, (it just forces the config to make it RE-compataible even if the test isn't compatible with RE).

Instead, we can set a `.buckconfig` option that will specify whether or not local runs are enabled. It will be on by default, and then on Sandcastle we will have to set it to `False` so that we always run using RE.

aniketmathur - apologies, but we'll have to add `-c android.buck2_android_instrumentation_test_can_run_locally=False` on our Sandcastle runs (we can also remove the `--unstable-force-tests-on-re`, that's won't actually be needed or do anything after this change). Thanks!

Reviewed By: aniketmathur

Differential Revision: D43351957

fbshipit-source-id: d8d21b8ce1e280c02d10200a82918703e5966c43