buck2: caching executor: give it its own RE Use Case / RE Properties
Summary:
Continuation of the previous diff. We need this to allow configuring a
caching executor without actual execution on RE. I don't think we intend to
use this, but it's a footgun we hit many times (when flipping `remote_enabled`
causes you to get cache misses), so it seems worth just getting rid of it
now that it's pretty easy.

Reviewed By: ndmitchell

Differential Revision: D43349778

fbshipit-source-id: 877238926f0ef0d455d2c4299ebb0fc4749fb8cc