buck2: OSS: RE GRPC: simplify client construction + use the right addresses
Summary:
Right now this uses only one address, which isn't right.

While fixing this, I started having to match on a bunch of things just to get
at the fields I wanted, so I also just redid our construction so that we don't
pretend to construct a Meta-internal RE config in the OSS build, that's just
way too complex and not useful.

Reviewed By: ndmitchell

Differential Revision: D43356570

fbshipit-source-id: 5cb8eb8aed9e91e8eb548d8c328047d281d0f6bf